Maybe `arch/common` was a bit overboard, things get a bit easier if the default-implementation macros reside next to the traits.
