--- src/arch/common/intrin/endian.rs
@@ -1,24 +1,0 @@
-use crate::arch::current::vecs::*;
-use crate::intrin::transmute::*;
-use crate::std::mem::transmute;
-
-macro_rules! impl_packed_swap_bytes {
-    ($vec:tt, $uvec:tt, $feat:expr, $mmfn:tt, ($($c:expr),*), ($($a:expr, $b:expr),*)) => {
-        impl Reendianize for $vec {
-            #[cfg(not(target_feature = $feat))]
-            #[inline(always)]
-            fn swap_bytes(&self) -> Self {
-                $vec::new($(self.extract($a).swap_bytes(),
-                            self.extract($b).swap_bytes()),*)
-            }
-
-            #[cfg(target_feature = $feat)]
-            #[inline(always)]
-            fn swap_bytes(&self) -> Self {
-                unsafe {
-                    transmute($mmfn(self.be_i8s(), $uvec::new($($c),*).be_i8s()))
-                }
-            }
-        }
-    }
-}
