--- src/arch/common/intrin/mod.rs
@@ -1,10 +1,0 @@
-#[macro_use] mod cast;
-#[macro_use] mod destride;
-#[macro_use] mod endian;
-#[macro_use] mod eq;
-#[macro_use] pub mod macros;
-#[macro_use] pub mod merge;
-#[macro_use] pub mod popcnt;
-#[macro_use] pub mod sum;
-#[macro_use] pub mod swizzle;
-#[macro_use] pub mod transmute;
