--- src/arch/common/mod.rs
@@ -1,2 +1,0 @@
-#[macro_use] pub mod intrin;
-#[macro_use] pub mod vecs;
