+++ src/intrin/macros.rs
