+++ src/intrin/recip.rs
@@ -5,6 +5,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
+
 pub trait Recip {
     /// Return a vector containing an estimation of the reciprocal of the
     /// corresponding elements of `self`.
