+++ src/intrin/sqrt.rs
@@ -21,3 +21,4 @@ pub trait Sqrt {
     /// ```
     fn sqrt(&self) -> Self;
 }
+
